AkiSky 1.13 登录页面改动，优化布局

+++ src/page/LoginPage.vue
@@ -1,30 +1,25 @@
 <template>
-  <div>
-    <div class="loginImg">
-      <img src="static/image/akiSkyHeader.png">
-    </div>
-    <div class="auth_left">
-      <div class="card">
-        <div class="card-body">
-          <form class="text-center" action="/login" method="post">
-            <div class="loginTitle">请登录您的账号</div>
-            <p v-if="tips!==''" style="text-align: center ;color: red">{{tips}}</p>
-            <el-input name="username" v-model="username" class="loginInput" placeholder="帐号" clearable></el-input>
-            <label class="form-label">
-            </label>
-            <el-input name="password" v-model="password" class="loginInput" placeholder="密码" show-password></el-input>
-<!--                 <label class="custom-control custom-checkbox">-->
-<!--                   <input type="checkbox" name="remember" class="custom-control-input">-->
-<!--                   <span class="custom-control-label">记住我</span>-->
-<!--                 </label>-->
-            <div style="text-align: center">
-              <el-button @click="submitLogin" type="primary">登录</el-button>
-            </div>
-          </form>
-        </div>
+  <div class="loginPage">
+    <div class="card">
+
+        <el-form ref="loginFormRef" :model="loginForm" :rules="rules" class="text-center" action="/login" method="post">
+          <div class="loginImg">
+            <img src="static/image/akiSkyHeader.png">
+          </div>
+<!--          <div class="loginTitle">请登录您的账号</div>-->
+<!--           <p v-if="tips!==''" style="text-align: center ;color: red">{{tips}}</p>-->
+          <el-form-item prop="username">
+            <el-input name="username" v-model="loginForm.username" placeholder="帐号" clearable></el-input>
+          </el-form-item>
+          <el-form-item prop="password">
+            <el-input name="password" v-model="loginForm.password" placeholder="密码" show-password></el-input>
+          </el-form-item>
+          <div style="text-align: center">
+            <el-button @click="submitLogin" type="primary">登录</el-button>
+          </div>
+        </el-form>
       </div>
     </div>
-  </div>
 </template>
 
 <script>
@@ -34,30 +29,43 @@ export default {
   name: "LoginPage",
   data(){
     return {
-      username: '',
-      password: '',
-      tips: ''
+      loginForm:{
+        username: '',
+        password: '',
+      },
+      rules:{
+        username: [
+          { required: true, message: '请输入帐号', trigger: 'blur' }
+        ],
+        password: [
+          { required: true, message: '请输入密码', trigger: 'blur' }
+        ]
+      }
     };
   },
   methods: {
     submitLogin(){
-      const params = new URLSearchParams();
-      params.append('username', this.username);
-      params.append('password', this.password);
-      axios.post('/login', params.toString())
-        .then(res=>{
-          if (res===401){
-            this.tips = '账号或密码错误！'
-            sessionStorage.clear()
-          }else if(res.code===200){
-            console.log(res.code)
-            this.tips = ''
-            this.DialogVisible = false
-            sessionStorage.setItem("name",res.data.name)
-            sessionStorage.setItem("avatarURL",res.data.avatarURL)
-            this.$router.push({path:'/Dynamic',query:{setid:123456}});
-          }
-        })
+      this.$refs.loginFormRef.validate(valid => {
+        if (!valid) return;
+        const params = new URLSearchParams();
+        params.append('username', this.loginForm.username);
+        params.append('password', this.loginForm.password);
+        axios.post('/login', params.toString())
+          .then(res=>{
+            if (res===401){
+              this.tips = '账号或密码错误！'
+              sessionStorage.clear()
+
+            }else if(res.code===200){
+              console.log(res.code)
+              this.tips = ''
+              this.DialogVisible = false
+              sessionStorage.setItem("name",res.data.name)
+              sessionStorage.setItem("avatarURL",res.data.avatarURL)
+              this.$router.push({path:'/Dynamic',query:{setid:123456}});
+            }
+          })
+      })
     },
     PopUp(){
       this.DialogVisible = true
@@ -68,42 +76,31 @@ export default {
 </script>
 
 <style scoped>
-.auth_left{
-  text-align: left;
-  box-sizing: border-box;
+.loginPage{
+  background: #a2cbf4;
+  height: 100%;
   display: flex;
+  justify-content: center;
   align-items: center;
-  height: 70vh;
-  width: 400px;
-  padding: 0 20px;
-  margin: 0 auto;
 }
 .card{
-  text-align: left;
-  direction: ltr;
+  width: 450px;
+  height: 330px;
   display: flex;
-  min-width: 0;
-  word-wrap: break-word;
+  justify-content: center;
+  align-items: center;
   background-color: #f6f8fa;
-  background-clip: border-box;
-  margin-bottom: 15px;
-  transition: all 0.5s ease-in-out;
-  border-radius: .55rem;
-  position: relative;
-  width: 100%;
+  border-radius: 20px;
 }
 .loginImg{
-  text-align: center;
+
 }
 .loginTitle{
   text-align: center;
   margin: 20px;
 }
-.loginInput{
-  width: 80%;
-  left: 10%;
-  margin-top: 10px;
-  margin-bottom: 20px;
+.el-input{
+
 }
 .el-button{
   margin-bottom: 30px;
